validation: move Retry func into utils (#712)
