+++ validation/utils.go
@@ -2,12 +2,14 @@ package validation
 
 import (
 	"bytes"
+	"context"
 	"fmt"
 	"math/big"
 	"os/exec"
 	"strings"
 	"testing"
 
+	"github.com/ethereum-optimism/optimism/op-service/retry"
 	"github.com/ethereum-optimism/superchain-registry/superchain"
 	"github.com/stretchr/testify/assert"
 )
@@ -55,4 +57,14 @@ func CastCall(contractAddress superchain.Address, calldata string, args []string
 	}
 
 	return results, nil
+}
+
+const DefaultMaxRetries = 3
+
+func Retry[S, T any](fn func(S) (T, error)) func(S) (T, error) {
+	return func(s S) (T, error) {
+		return retry.Do(context.Background(), DefaultMaxRetries, retry.Exponential(), func() (T, error) {
+			return fn(s)
+		})
+	}
 }
